inference-backend/requirements.txt
Flask
deepface
pillow
numpy
tf-keras
scikit-learn
opencv-python
joblib
ffmpeg-python